Less confusing use of temperature units
The ISY defines units for both C and F so lets use them. The
sensibo tells us what units it is using so use the same for
the node values.

The set temperature is a bit different in that the selection list
will contain both so the user can pick which they want to use to
set the target temperature.